stt_service/pyproject.toml
[tool.poetry]
name = "operator-voice-trainer-stt"
version = "0.1.0"
description = "STT service for Operator Voice Trainer"
authors = ["sssay10 <duvakin.k@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.122.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = "^2.0.0"
# STT dependencies - OpenAI Whisper
openai-whisper = {git = "https://github.com/openai/whisper.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

